Add sphinx_toml (future IPython docs building dependency).
I'm planning to start using this as a dependency for IPython when
building docs, in order to avoid as much as possible to have config in a
conf.py file ... that is until sphinx support static configuration file.

Of course; as IPython is released regularly I would prefer to have
that ready in conda_forge before and having the possibility to making
a new release quickly before making it a dependency in IPython.